+++ gazeplay-games/src/main/java/net/gazeplay/games/creampie/CreampieGameLauncher.java
@@ -24,7 +24,7 @@ public class CreampieGameLauncher implements IGameLauncher<Stats, IGameVariant> 
                                   RoundsDurationReport roundsDurationReport,
                                   List<List<FixationPoint>> fixationSequence,
                                   List<CoordinatesTracker> movementHistory,
-                                  double[][] heatMap,
+                                  int[][] heatMap,
                                   List<AreaOfInterest> aoiList,
                                   SavedStatsInfo savedStatsInfo
     ) {
