+++ gazeplay-games/src/main/java/net/gazeplay/games/math101/MathGameLauncher.java
@@ -28,7 +28,7 @@ public final class MathGameLauncher implements IGameLauncher<Stats, MathGameVari
                                   RoundsDurationReport roundsDurationReport,
                                   List<List<FixationPoint>> fixationSequence,
                                   List<CoordinatesTracker> movementHistory,
-                                  double[][] heatMap,
+                                  int[][] heatMap,
                                   List<AreaOfInterest> aoiList,
                                   SavedStatsInfo savedStatsInfo
     ) {
