+++ gazeplay-games/src/main/java/net/gazeplay/games/slidingpuzzle/SlidingPuzzleStats.java
@@ -27,7 +27,7 @@ class SlidingPuzzleStats extends SelectionGamesStats {
                        RoundsDurationReport roundsDurationReport,
                        List<List<FixationPoint>> fixationSequence,
                        List<CoordinatesTracker> movementHistory,
-                       double[][] heatMap,
+                       int[][] heatMap,
                        List<AreaOfInterest> aoiList,
                        SavedStatsInfo savedStatsInfo
     ) {
