+++ gazeplay-games/src/main/java/net/gazeplay/games/soundsoflife/SoundsOfLifeStats.java
@@ -23,7 +23,7 @@ public class SoundsOfLifeStats extends SelectionGamesStats {
                              RoundsDurationReport roundsDurationReport,
                              List<List<FixationPoint>> fixationSequence,
                              List<CoordinatesTracker> movementHistory,
-                             double[][] heatMap,
+                             int[][] heatMap,
                              List<AreaOfInterest> aoiList,
                              SavedStatsInfo savedStatsInfo
     ) {
